Added clang tidy support through dotfile and cmake.
Still need to clean up a few errors and tighten linting, but it looks a
lot better now. THis is a big step in the right direction.

Static analysis is crutial, and will be enforced for the remainder of
the projects work.

Will merge this into the main branch once I clean up static analysis and
add it to the workflow!

+++ tests/util/test_alphanum.cpp
@@ -1,7 +1,7 @@
 #include "catch_amalgamated.hpp"
 
-#include <stdint.h>
-#include <stdlib.h>
+#include <cstdint>
+#include <cstdlib>
 
 #include <vector>
 
@@ -132,13 +132,13 @@ TEST_CASE("Malformed numbers") {
     SECTION("Signing integers") {
         int64_t v = start;
         REQUIRE(strntoll("0b2", 3, Base::DECIMAL, &v) == Status::MALFORMED_INTEGER_STR);
-        REQUIRE(v == start);
+        REQUIRE(v == (int64_t)start);
         REQUIRE(strntoll("0b2", 3, Base::BINARY, &v) == Status::MALFORMED_INTEGER_STR);
-        REQUIRE(v == start);
+        REQUIRE(v == (int64_t)start);
         REQUIRE(strntoll("0xG", 3, Base::HEXADECIMAL, &v) == Status::MALFORMED_INTEGER_STR);
-        REQUIRE(v == start);
+        REQUIRE(v == (int64_t)start);
         REQUIRE(strntoll("0oG", 3, Base::OCTAL, &v) == Status::MALFORMED_INTEGER_STR);
-        REQUIRE(v == start);
+        REQUIRE(v == (int64_t)start);
     }
 
     SECTION("Unsigned integers") {
